updateFeatureActions2 working needs cleanup
